CR#24034 Fix and clean up tests
